+++ packages/hix/lib/Hix/Component.hs
@@ -2,7 +2,6 @@ module Hix.Component where
 
 import Control.Monad.Trans.Class (lift)
 import Control.Monad.Trans.Except (throwE)
-import Control.Monad.Trans.Reader (ask)
 import Data.List.Extra (firstJust)
 import qualified Data.Map.Strict as Map
 import Data.Map.Strict ((!?))
@@ -20,15 +19,15 @@ import Hix.Data.ComponentConfig (
   Target (Target),
   )
 import Hix.Data.Error (Error (EnvError), pathText)
-import qualified Hix.Monad as Monad
-import Hix.Monad (Env (Env), M, noteEnv)
+import Hix.Monad (M, noteEnv)
 import qualified Hix.Options as Options
 import Hix.Options (
   ComponentCoords,
   ComponentSpec (ComponentSpec),
   PackageSpec (PackageSpec),
-  TargetSpec (TargetDefault, TargetForComponent, TargetForFile),
+  TargetSpec (TargetForComponent, TargetForFile),
   )
+import Hix.Path (rootDir)
 
 tryPackageByDir ::
   PackagesConfig ->
@@ -46,13 +45,18 @@ packageByDir ::
 packageByDir config dir =
   noteEnv [exon|No package at this directory: #{pathText dir}|] (tryPackageByDir config dir)
 
+packageDefault :: PackagesConfig -> M PackageConfig
+packageDefault = \case
+  [(_, pkg)] -> pure pkg
+  _ -> lift (throwE (EnvError "Project has more than one package, specify -p or -f."))
+
 packageForSpec ::
+  Path Abs Dir ->
   PackagesConfig ->
   PackageSpec ->
   M PackageConfig
-packageForSpec config = \case
+packageForSpec root config = \case
   PackageSpec _ (Just (Abs dir)) -> do
-    Env {root} <- ask
     rel <- noteEnv [exon|Path is not a subdirectory of the project root: #{pathText dir}|] (stripProperPrefix root dir)
     packageByDir config rel
   PackageSpec (PackageName name) (Just (Rel dir)) | Text.elem '/' name ->
@@ -63,6 +67,15 @@ packageForSpec config = \case
       tryDir = \case
         Abs _ -> Nothing
         Rel rd -> tryPackageByDir config rd
+
+packageForSpecOrDefault ::
+  Path Abs Dir ->
+  PackagesConfig ->
+  Maybe PackageSpec ->
+  M PackageConfig
+packageForSpecOrDefault root config = \case
+  Just pkg -> packageForSpec root config pkg
+  Nothing -> packageDefault config
 
 matchComponent :: ComponentConfig -> ComponentSpec -> Bool
 matchComponent candidate (ComponentSpec name dir) =
@@ -101,19 +114,20 @@ targetInPackage package = \case
       match = flip matchComponent
 
 targetForComponent ::
+  Path Abs Dir ->
   PackagesConfig ->
   ComponentCoords ->
   M Target
-targetForComponent config spec = do
-  package <- packageForSpec config spec.package
+targetForComponent root config spec = do
+  package <- packageForSpecOrDefault root config spec.package
   targetInPackage package spec.component
 
 targetForFile ::
+  Path Abs Dir ->
   PackagesConfig ->
   Path Abs File ->
   M Target
-targetForFile config file = do
-  Env {root} <- ask
+targetForFile root config file = do
   fileRel <- stripProperPrefix root file
   (package, subpath) <- pkgError (firstJust (matchPackage fileRel) (Map.elems config))
   (component, sourceDir) <- compError (firstJust (matchSourceDir subpath) (Map.elems package.components))
@@ -129,20 +143,22 @@ targetForFile config file = do
     pkgError = noteEnv "No package contains this file"
     compError = noteEnv "No component source dir contains this file"
 
-targetDefault :: PackagesConfig -> Maybe ComponentSpec -> M Target
-targetDefault [(_, pkg)] comp =
-  targetInPackage pkg comp
-targetDefault _ _ =
-  lift (throwE (EnvError "Project has more than one package, specify -p or -f."))
-
-targetComponent ::
+targetComponentIn ::
+  Path Abs Dir ->
   PackagesConfig ->
   TargetSpec ->
   M Target
-targetComponent config = \case
+targetComponentIn root config = \case
   TargetForComponent spec ->
-    targetForComponent config spec
+    targetForComponent root config spec
   TargetForFile spec ->
-    targetForFile config spec
-  TargetDefault spec ->
-    targetDefault config spec
+    targetForFile root config spec
+
+targetComponent ::
+  Maybe (Path Abs Dir) ->
+  PackagesConfig ->
+  TargetSpec ->
+  M Target
+targetComponent cliRoot config spec = do
+  root <- rootDir cliRoot
+  targetComponentIn root config spec